Typography audit
The only text within the overlay itself is just the title - it should be styled using d2l-typography. Despite it having the class previously, it wasn't actually being applied - it was just the two "overrides" that were styling it. Amusingly, those overrides are the actual values if it's applied properly. hah.